chore: cleanup code and comment
