+++ src/HOCs/index.js
@@ -1,0 +1,2 @@
+export * from "./deprecatePropsHOC";
+export * from "./shouldComponentUpdateHOC";
